// src/common/enum.ts
export enum FieldTypeEnum {
    inputText = 'Input',
    textarea = 'Textarea',
    number = 'InputNumber',
    select = 'Select',
    treeSelect = 'TreeSelect',
    datePicker = 'DatePicker',
    dateRange = 'DateRange',
    switch = 'Switch',
    radio = 'Radio',
    checkbox = 'Checkbox',
    file = 'File',
    staticText = 'StaticText',
}

export enum ValidationTypeEnum {
    required = 'required',
    pattern = 'pattern',
    validator = 'validator',
}